Move close/toggle pane actions to separate files
